fix test for language detect
